Moved sync into main folder
